Refactor code to send complete event and use PUT in push of observations
